requirements.txt
black==19.3b0
ruamel.yaml
yamllint
